Added documentation for the Kafka egress connector.
