+++ server/swimos_connector_kafka/src/deser/mod.rs
@@ -71,7 +71,7 @@ pub enum MessagePart {
     Payload,
 }
 
-/// A deserializer that will attempt to produce a [value](Value) from a component of a Kafka messsage.
+/// A deserializer that will attempt to produce a [value](Value) from a component of a Kafka message.
 pub trait MessageDeserializer {
     type Error: std::error::Error;
